% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beijing2022}
\alias{beijing2022}
\title{Title}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3785 rows and 7 columns.
}
\usage{
beijing2022
}
\description{
Title
}
\details{
TODO
}
\examples{
beijing2022
}
\keyword{datasets}
