% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beijing_events}
\alias{beijing_events}
\title{Title}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 109 rows and 4 columns.
}
\usage{
beijing_events
}
\description{
Title
}
\details{
TODO
}
\examples{
beijing_events
}
\keyword{datasets}
