% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{get_events}
\alias{get_events}
\title{Title}
\usage{
get_events(game, sport)
}
\arguments{
\item{game}{the olympic game, in the format of tokyo-2020, see the slug column in olympic_game}

\item{sport}{the sport, see the slug column from get_sports()}
}
\value{
a tibble of event summary
}
\description{
Title
}
\examples{
get_events(game = "tokyo-2020", sport = "swimming")
}
