% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medals.R
\name{get_medals}
\alias{get_medals}
\title{Title}
\usage{
get_medals(game)
}
\arguments{
\item{game}{the olympic game, in the format of tokyo-2020, see the slug column in olympic_game}
}
\value{
a tibble of medal table
}
\description{
Title
}
\examples{
get_medals(game = "tokyo-2020")
}
