% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{get_results}
\alias{get_results}
\title{Title}
\usage{
get_results(url)
}
\arguments{
\item{url}{the url to event page in olympic.com, i.e. https://olympics.com/en/olympic-games/tokyo-2020/results/swimming/men-s-100m-breaststroke}
}
\value{
a tibble of result summary
}
\description{
Title
}
\examples{
url <- "https://olympics.com/en/olympic-games/tokyo-2020/results/swimming/men-s-100m-breaststroke"
get_results(url)
}
