% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{get_single_event}
\alias{get_single_event}
\title{Title}
\usage{
get_single_event(url)
}
\arguments{
\item{url}{the URL to the event}
}
\value{
a tibble of event summary
}
\description{
Title
}
\examples{
get_single_event("https://olympics.com/en/olympic-games/tokyo-2020/results/shooting")
}
