% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{get_sports}
\alias{get_sports}
\title{Title}
\usage{
get_sports(game)
}
\arguments{
\item{game}{the olympic game, in the format of tokyo-2020, see the slug column in olympic_game}
}
\value{
a tibble of event summary
}
\description{
Title
}
\examples{
get_sports(game = "tokyo-2020")
}
