% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medals}
\alias{medals}
\title{Title}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1747 rows and 6 columns.
}
\usage{
medals
}
\description{
Title
}
\details{
TODO
}
\examples{
medals
}
\keyword{datasets}
