% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{olympic_game}
\alias{olympic_game}
\title{Title}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 53 rows and 7 columns.
}
\usage{
olympic_game
}
\description{
Title
}
\details{
TODO
}
\examples{
olympic_game
}
\keyword{datasets}
