% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{get_sports}
\alias{get_sports}
\alias{get_single_event}
\alias{get_events}
\alias{get_single_result}
\alias{get_results}
\title{Functions to scrape olympic data}
\usage{
get_sports(game)

get_single_event(url)

get_events(table)

get_single_result(url)

get_results(table)
}
\arguments{
\item{game}{used in \code{\link[=get_sports]{get_sports()}}; \cr in the format of tokyo-2020, see the slug column in the built-in data \link{olympic_game}.}

\item{url}{used in \code{\link[=get_single_event]{get_single_event()}} and \code{\link[=get_single_result]{get_single_result()}}; \cr the URL to the sport/ event page, see examples.}

\item{table}{used in \code{\link[=get_events]{get_events()}} and \code{\link[=get_results]{get_results()}}; \cr the tibble produced by \code{\link[=get_sports]{get_sports()}} and \code{\link[=get_events]{get_events()}}, see examples.}
}
\value{
a tibble of event summary
}
\description{
Functions to scrape olympic data from olympic.com
}
\examples{
# sport table in Tokyo 2020
get_sports(game = "tokyo-2020")

# event table from URL or sport table
get_single_event("https://olympics.com/en/olympic-games/tokyo-2020/results/swimming")
get_sports(game = "tokyo-2020") \%>\% head(1) \%>\% get_events()

# event results from URL or event table
url <- "https://olympics.com/en/olympic-games/tokyo-2020/results/swimming/women-s-800m-freestyle"
get_single_result(url)
get_sports(game = "tokyo-2020") \%>\% head(1) \%>\% get_events() \%>\% head(1) \%>\% get_results()
}
