% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tokyo2020}
\alias{tokyo2020}
\title{Title}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6900 rows and 7 columns.
}
\usage{
tokyo2020
}
\description{
Title
}
\details{
TODO
}
\examples{
tokyo2020
}
\keyword{datasets}
