% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{winter_games}
\alias{winter_games}
\title{Title}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17711 rows and 7 columns.
}
\usage{
winter_games
}
\description{
Title
}
\details{
TODO
}
\examples{
winter_games
}
\keyword{datasets}
