% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medals.R
\name{get_medals}
\alias{get_medals}
\title{Scrape Olympic medal board}
\usage{
get_medals(game)
}
\arguments{
\item{game}{the Olympic game, in the format of tokyo-2020, see the slug column in the built-in olympic_game data}
}
\value{
a tibble of medal table
}
\description{
Scrape Olympic medal board
}
\examples{
get_medals(game = "tokyo-2020")
}
