% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{get_results}
\alias{get_results}
\title{Title}
\usage{
get_results(table)
}
\arguments{
\item{table}{the output from get_event}
}
\value{
a tibble of result summary
}
\description{
Title
}
\examples{
get_sports(game = "tokyo-2020") \%>\% head(1) \%>\% get_events() \%>\% head(1) \%>\% get_results()
}
