% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olympic-package.R
\docType{package}
\name{olympics-package}
\alias{olympics-package}
\alias{_PACKAGE}
\alias{olympic-package}
\title{olympics: Olympic data from olympic.com}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An interface to scrape olympic results from olympic.com and built-in data of 21st century summer and winter game results.
}
\author{
\strong{Maintainer}: Shery Zhang \email{huize.zhang@monash.edu}

Authors:
\itemize{
  \item Xiaoqian Wang \email{Xiaoqian.Wang@monash.edu}
  \item Nuwani Palihawadana \email{Nuwani.KodikaraPalihawadana@monash.edu}
}

}
\keyword{internal}
